Organized server and client
