+++ README.md
@@ -1,0 +1,8 @@
+cd client
+npm install
+
+cd ../server
+npm install
+
+cd ..
+npm start
