+++ config/dev.js
@@ -1,0 +1,3 @@
+module.exports = {
+  mongoURI: "mongodb://dev:dev@ds149324.mlab.com:49324/mobadraft-dev"
+};
