+++ server/models/HotsDraft.js
