+++ server/package.json
@@ -1,0 +1,26 @@
+{
+  "name": "mobadraft",
+  "version": "1.0.0",
+  "description":
+    "A web application dedicated to the drafting element of Moba PC games such as Heroes of The Storm, Dota 2, and League of Legends. This Will be used to improve the drafting element of players, amateur teams, and professional players.",
+  "main": "index.js",
+  "engines": {
+    "node": "8.1.2",
+    "npm": "5.4.2"
+  },
+  "scripts": {
+    "start": "nodemon index.js"
+  },
+  "author": "Nicholas March",
+  "license": "ISC",
+  "dependencies": {
+    "body-parser": "^1.18.2",
+    "express": "^4.16.2",
+    "moment": "^2.18.1",
+    "mongoose": "^4.11.14",
+    "nodemon": "^1.12.1",
+    "randomstring": "^1.1.5",
+    "socket.io": "^2.0.3",
+    "winston": "^2.3.1"
+  }
+}
