+++ server/routes/hotsRoutes.js
